+++ lib/Auth/Process/OpaqueSmartID.php
@@ -90,9 +90,14 @@
 class sspmod_userid_Auth_Process_OpaqueSmartID extends SimpleSAML_Auth_ProcessingFilter {
 
     /**
-	 * List of tags that the auth process should be executed
-	 */
+     * List of tags that the auth process should be executed
+     */
     private $idpTagWhitelist = array();
+
+    /**
+     * List of tags that the auth process should not be executed
+     */
+    private $idpTagBlacklist = array();
 
     // List of IdP entityIDs that should be excluded from the authority
     // part of the user id source.
@@ -149,6 +154,13 @@ class sspmod_userid_Auth_Process_OpaqueSmartID extends SimpleSAML_Auth_Processin
             $this->idpTagWhitelist = $config['idp_tag_whitelist'];
             if (!is_array($this->idpTagWhitelist)) {
                 throw new Exception('OpaqueSmartID authproc configuration error: \'idp_tag_whitelist\' should be an array.');
+            }
+        }
+
+        if (array_key_exists('idp_tag_blacklist', $config)) {
+            $this->idpTagBlacklist = $config['idp_tag_blacklist'];
+            if (!is_array($this->idpTagBlacklist)) {
+                throw new Exception('OpaqueSmartID authproc configuration error: \'idp_tag_blacklist\' should be an array.');
             }
         }
 
@@ -212,14 +224,25 @@ class sspmod_userid_Auth_Process_OpaqueSmartID extends SimpleSAML_Auth_Processin
         assert('array_key_exists("Attributes", $request)');
 
         $idpMetadata = $this->getIdPMetadata($request);
-		$idpTags = $this->getIdPTags($idpMetadata);
-        if (empty(array_intersect($this->idpTagWhitelist, $idpTags))) {
+        $idpTags = $this->getIdPTags($idpMetadata);
+        if (!empty(array_intersect($this->idpTagBlacklist, $idpTags))) {
             if ($this->setUserIdAttribute && !empty($request['Attributes'][$this->idAttribute])) {
                 $request['UserID'] = $request['Attributes'][$this->idAttribute][0];
             }
             SimpleSAML\Logger::debug("[OpaqueSmartID] Skipping IdP with tags " . var_export($idpTags, true));
             return;
+        } else {
+            if (!empty($this->idpTagWhitelist)) {
+                if (empty(array_intersect($this->idpTagWhitelist, $idpTags))) {
+                    if ($this->setUserIdAttribute && !empty($request['Attributes'][$this->idAttribute])) {
+                        $request['UserID'] = $request['Attributes'][$this->idAttribute][0];
+                    }
+                    SimpleSAML\Logger::debug("[OpaqueSmartID] Skipping IdP with tags " . var_export($idpTags, true));
+                    return;
+                }
+            }
         }
+
         $userId = $this->generateUserId($request['Attributes'], $request);
 
         if (isset($userId)) {